+++ src/devTools/editor/hooks/usePipelineField.ts
@@ -27,11 +27,11 @@ import validateRenderers from "@/devTools/editor/validation/validateRenderers";
 import applyTraceError from "@/devTools/editor/validation/applyTraceError";
 import { isEmpty } from "lodash";
 import {
-  BlocksMap,
   FormikError,
   FormikErrorTree,
 } from "@/devTools/editor/tabs/editTab/editTabTypes";
 import { ElementType } from "@/devTools/editor/extensionPoints/elementConfig";
+import { BlocksMap } from "@/blocks/registry";
 
 export const PIPELINE_BLOCKS_FIELD_NAME = "extension.blockPipeline";
